# openapi.yaml
openapi: 3.0.3
info:
  title: AI 食物卡路里辨識系統 API
  description: |
    一套完整的 AI 識別食物卡路里 Web 應用程式 API
    包含會員系統、拍照辨識、飲食記錄、數據分析等功能
  version: 1.0.0
  contact:
    name: API Support
    email: support@calorie-ai.com
servers:
  - url: https://api.calorie-ai.com/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
...
  /auth/register:
    post:
      tags:
        - Auth
      summary: 會員註冊
      description: 使用 Email 註冊新帳號
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: SecurePass123!
                name:
                  type: string
                  example: 王小明
                dateOfBirth:
                  type: string
                  format: date
                  example: 1990-01-01
                gender:
                  type: string
                  enum: [male, female, other]
                  example: male
                height:
                  type: number
                  description: 身高(公分)
                  example: 170
                weight:
                  type: number
                  description: 體重(公斤)
                  example: 65
                targetWeight:
                  type: number
                  description: 目標體重(公斤)
                  example: 60
      responses:
        '201':
          description: 註冊成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 註冊成功，請查收驗證郵件
                  data:
                    type: object
                    properties:
                      userId:
                        type: string
                        example: usr_1234567890
                      email:
                        type: string
                        example: user@example.com
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email 已被註冊
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/forgot-password:
    post:
      tags:
        - Auth
      summary: 忘記密碼
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: 重置郵件已發送
  /auth/reset-password:
    post:
      tags:
        - Auth
      summary: 重置密碼
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - newPassword
              properties:
                token:
                  type: string
                newPassword:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: 密碼重置成功

  # ==================== 會員管理 ====================
...
  /recognition/upload:
    post:
      tags:
        - Food Recognition
      summary: 上傳圖片進行食物辨識
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
                  description: 食物圖片 (支援 JPG, PNG, HEIC)
                mealType:
                  type: string
                  enum: [breakfast, lunch, dinner, snack]
                  description: 餐點類型
                timestamp:
                  type: string
                  format: date-time
                  description: 拍攝時間
      responses:
        '200':
          description: 辨識成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      recognitionId:
                        type: string
                        example: rec_1234567890
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/RecognizedFood'
                      totalCalories:
                        type: number
                        example: 520
                      totalNutrition:
                        $ref: '#/components/schemas/Nutrition'
                      imageUrl:
                        type: string
                        format: uri
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          description: 圖片檔案過大
  /recognition/{recognitionId}:
    get:
      tags:
        - Food Recognition
      summary: 取得辨識結果
      security:
        - BearerAuth: []
      parameters:
        - name: recognitionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功取得辨識結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecognitionResult'
    
    patch:
      tags:
        - Food Recognition
      summary: 修正辨識結果
      security:
        - BearerAuth: []
      parameters:
        - name: recognitionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      itemId:
                        type: string
                      foodName:
                        type: string
                      servingSize:
                        type: number
                      calories:
                        type: number
                      nutrition:
                        $ref: '#/components/schemas/Nutrition'
      responses:
        '200':
          description: 修正成功
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Nutrition:
      type: object
      properties:
        calories:
          type: number
          example: 250
        protein:
          type: number
          example: 15.5
        carbs:
          type: number
          example: 30.2
        fat:
          type: number
          example: 8.3
        fiber:
          type: number
          example: 3.5
        sugar:
          type: number
          example: 5.2
        sodium:
          type: number
          example: 350
    RecognizedFood:
      type: object
      properties:
        itemId:
          type: string
        foodId:
          type: string
        foodName:
          type: string
          example: 白飯
        confidence:
          type: number
          description: 辨識信心度 (0-1)
          example: 0.95
        servingSize:
          type: number
          example: 150
        unit:
          type: string
          example: 克
        calories:
          type: number
          example: 210
        nutrition:
          $ref: '#/components/schemas/Nutrition'
        boundingBox:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
            width:
              type: number
            height:
              type: number
    RecognitionResult:
      type: object
      properties:
        recognitionId:
          type: string
        imageUrl:
          type: string
          format: uri
        items:
          type: array
          items:
            $ref: '#/components/schemas/RecognizedFood'
        totalCalories:
          type: number
        totalNutrition:
          $ref: '#/components/schemas/Nutrition'
        createdAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: 請求參數有誤
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
  responses:
    BadRequest:
      description: 請求參數錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...